+++ CHANGES.md
@@ -96,3 +96,31 @@ v0.83.0
 * feat: closing the window closes the application (as for FCPX...) which is not the way MacOS works.
 * fix: the version of the software is now displayed in the title bar.
 * feat: the console tool does appear only in debug mode.
+
+v0.85.0
+-------
+
+* fix: display of lost files (React issue).
+* fix: opening the FCPX library via the button will not scroll up the screen anymore.
+* fix: in the tab "Backups", data could be totally incorrect under some circumstances.
+* feat: the backup store file is saved once modified (when a library is added to the backup list).
+* fix: when a library is moved, then the backup store reflect silently the change.
+
+v0.90.0
+-------
+
+* fix: issue displaying libraries as duplicated if already backuped.
+* feat: keep the power up during the running of the software.
+* refactor: use of &lt;Spinner&gt; from React Bootstrap.
+
+v0.95.0
+-------
+
+* feat: the scan of the libraries and their backup is now sequential (esier for the end user).
+* fix: issue in saving the store.json
+* refactor: the way the refresh function works. Now the fcpx-renderer code is responsible for synchronisation.
+* refactor: the scan is now fully asynchronously and made _before_ the scan of each library.
+* refactor: the async/Promises are now well understood and well integrated.
+* feat: added a logo for the application in the main screen and reorganized display
+* feat: if the backup disk is not inserted, the display shows an alert and the backup tab is hidden.
+